fix up moved imports/renamed methods
