+++ src/tests/end-to-end/common/browser.ts
@@ -4,9 +4,9 @@ import * as Puppeteer from 'puppeteer';
 import { browserLogPath } from './browser-factory';
 import { popupPageElementIdentifiers } from './element-identifiers/popup-page-element-identifiers';
 import { forceTestFailure } from './force-test-failure';
-import { Page } from './page';
 import { BackgroundPage, isBackgroundPageTarget } from './page-controllers/background-page';
 import { DetailsViewPage } from './page-controllers/details-view-page';
+import { Page } from './page-controllers/page';
 import { PopupPage } from './page-controllers/popup-page';
 import { TargetPage } from './page-controllers/target-page';
 import { getTestResourceUrl } from './test-resources';
