+++ src/tests/end-to-end/common/page-controllers/page.ts
@@ -3,9 +3,9 @@
 import * as Puppeteer from 'puppeteer';
 
 import { includes } from 'lodash';
-import { forceTestFailure } from './force-test-failure';
-import { takeScreenshot } from './generate-screenshot';
-import { DEFAULT_NEW_PAGE_WAIT_TIMEOUT_MS, DEFAULT_PAGE_ELEMENT_WAIT_TIMEOUT_MS } from './timeouts';
+import { forceTestFailure } from '../force-test-failure';
+import { takeScreenshot } from '../generate-screenshot';
+import { DEFAULT_NEW_PAGE_WAIT_TIMEOUT_MS, DEFAULT_PAGE_ELEMENT_WAIT_TIMEOUT_MS } from '../timeouts';
 
 export type PageOptions = {
     onPageCrash?: () => void;
