+++ src/tests/end-to-end/common/page-controllers/popup-page.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import * as Puppeteer from 'puppeteer';
-import { Page } from '../page';
+import { Page, PageOptions } from './page';
 
 export class PopupPage extends Page {
     constructor(underlyingPage: Puppeteer.Page, options?: PageOptions) {
